FROMLIST: drm/panel: samsung-sofef00-bonito: move dsi commands to enable function
DSI commands are no longer allowed in the prepare function. Move it to a
new enable function so that the panel can successfully turn on.

This is not a confirmed bug on this panel, but was discovered on the
S6E3FA7 panel.

Signed-off-by: Richard Acayan <mailingradian@gmail.com>

(cherry picked from https://gitlab.com/sdm670-mainline/linux-patches stable/drm-panel-samsung-sofef00-bonito-move-dsi-commands-to-enable-function.patch @ b306510dd9b960c9e030ea9973c14a81890120b3)